alteracao do axios e modularizacao do app feitas
